Avoid wrong node grouping when nodes start from similar substring
Some environments have nodes that start from the similar substring,
e.g. ceph001 and ceph-mon01. For this example, the node name is
parsed as 'ceph' and 'ceph-mon' correctly. But 'ceph' pcre is
used for gathering OSD nodes where ceph001 and ceph-mon001 are
grouped togather by mistake. To avoid this, suggest to search with
more precise reg.exp.