Fixed K8S Node status test
The node status in K8S cluster can be "Ready", "NotReady", etc.
If we check the substring "Ready" in "NotReady" status, it can
be found and the test passes. Need to check the entire word "Ready"
in the status field.

Example: test failed on my env:
AssertionError: k8s is not healthy:
      "cmp024    NotReady   node      103d      v1.8.5-4+77af14b530a816",
      "cmp046    NotReady   node      103d      v1.8.5-4+77af14b530a816"